+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add `threeDSecureAuthenticationId` support on transaction sale
 * Add ThreeDSecure test payment method nonces
 * Add test `AuthenticationId`s Unreleased
+* Add `DisputeAccepted`, `DisputeDisputed`, and `DisputeExpired` webhook constants
 
 ## 2.22.0
 * Add `RefundAuthHardDeclined` and `RefundAuthSoftDeclined` to validation errors